Made filepaths relative, added variables to config distributions
For io in capitals, I changed to ..inputs/housing_stock.csv so it works
on all machines.

For config.py I created variables for each mean and standard deviation
and put references to them in each guassian function to make it easier
to change values. I set all std to zero to take out stockasticity.